\name{damitools-package}
\alias{damitools-package}
\alias{damitools}
\docType{package}
\title{
Collection of R Tools Written by Damiano Fantini
}
\description{
Collection of tools and Functions written for Data Science and Bioinformatics projects.
}

\author{
Damiano Fantini <damiano.fantini@gmail.com>
}
